Move high level tests to bdd specifications
